+++ src/Plainion.CI.Redist/Properties/AssemblyInfo.cs
@@ -5,15 +5,15 @@ using System.Reflection;
 [assembly: AssemblyProductAttribute("Plainion.CI")]
 [assembly: AssemblyDescriptionAttribute("Plainion.CI")]
 [assembly: AssemblyCopyrightAttribute("Copyright @ 2017")]
-[assembly: AssemblyVersionAttribute("1.3")]
-[assembly: AssemblyFileVersionAttribute("1.3")]
+[assembly: AssemblyVersionAttribute("1.4")]
+[assembly: AssemblyFileVersionAttribute("1.4")]
 namespace System {
     internal static class AssemblyVersionInformation {
         internal const System.String AssemblyTitle = "Plainion.CI.Redist";
         internal const System.String AssemblyProduct = "Plainion.CI";
         internal const System.String AssemblyDescription = "Plainion.CI";
         internal const System.String AssemblyCopyright = "Copyright @ 2017";
-        internal const System.String AssemblyVersion = "1.3";
-        internal const System.String AssemblyFileVersion = "1.3";
+        internal const System.String AssemblyVersion = "1.4";
+        internal const System.String AssemblyFileVersion = "1.4";
     }
 }